refactor: split router into own objects for clarity
ref: #233 #227
